backend/requirements.txt
flask
psycopg2-binary  # Pour connecter Flask à PostgreSQL
